Change the backend to reflect the integer input

+++ backend/app/models/bucket_models.py
@@ -14,7 +14,7 @@ class Bucket(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
     name: Mapped[str] = mapped_column(String)
     description: Mapped[str] = mapped_column(String)
-    amount: Mapped[float] = mapped_column(Integer)
+    amount: Mapped[int] = mapped_column(Integer)
     variant: Mapped[dict] = mapped_column(DatetimeJSON)
     created_at: Mapped[datetime] = mapped_column(DateTime)
     updated_at: Mapped[datetime] = mapped_column(DateTime)
